RDKBDEV-3132: Add flag for including Telemetry's download_t2_profile.sh
Reason for change: T2 profiles are not downloaded automatically if
web config is not used which is inconvenient for developers.
Test procedures: Configure a clean build with envvar T2_WITHOUT_WEBCONFIG=y.
Build and deploy the image to a device with a valid T2 profile/rule configured in xconf.
Check that is present, and run it "/lib/rdk/download_t2_profile.sh".
Check that the report profile has been added to the data model.
Risks: Low

Change-Id: If15aecb5b5df19ceb5e17fe0f75e7854b861863f
Signed-off-by: Bryce Young <bryce.young@cognizant.com>